selectTaskAsListItem
