+++ src/selectors/tasks/select_saved_tasks.ts
@@ -1,7 +1,7 @@
 import { Store } from '../../stores';
 import * as store from '../../stores/tasks';
 import { TaskListItem } from './task_list_item';
-import { selectTaskAsListItem } from '.';
+import { selectTaskAsListItem } from './select_task_as_list_item';
 
 export const selectSavedTasks = (appStore: Store): ReadonlyArray<TaskListItem> => {
     const savedTasksList = appStore.tasksInStore.savedTasksList;
